pyproject.toml
[project]
name = "rv-search-agent"
version = "0.1.0"
description = "An AI agent for searching and analyzing RV listings"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.39.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
